final touch

+++ src/app/body/completed/completed.component.ts
@@ -3,6 +3,7 @@ import { CompletedService } from 'src/app/shared/services/completed.service';
 import { Task } from 'src/app/shared/task.model';
 import { CompletedStorageService } from 'src/app/shared/storage/completed-storage.service';
 import { TasksService } from 'src/app/shared/services/tasks.service';
+import { AuthService } from 'src/app/auth/auth.service';
 
 @Component({
   selector: 'app-completed',
@@ -11,9 +12,11 @@ import { TasksService } from 'src/app/shared/services/tasks.service';
 })
 export class CompletedComponent implements OnInit {
   completTasks: Task[];
+  isAdmin = false;
   constructor(private completedService: CompletedService,
               private completedStorage: CompletedStorageService,
-              private tasksService: TasksService) { }
+              private tasksService: TasksService,
+              private authService: AuthService) { }
 
   ngOnInit() {
     this.completedStorage.fetchTasks().subscribe()
@@ -24,6 +27,18 @@ export class CompletedComponent implements OnInit {
         this.completedStorage.storeCompletedTasks();
       }
     )
+    this.authService.user.subscribe(
+      user =>{
+        if(!user){
+          return
+        }
+        if(user.email === 'admin@test.com'){
+          this.isAdmin = true;          
+        } else {
+          this.isAdmin = false;
+        }
+      }
+    )
   }
   onSendTasks(index: number){
     const taskToSend = this.completedService.geTask(index);
@@ -31,6 +46,9 @@ export class CompletedComponent implements OnInit {
     this.onDelete(index);
   }
   onDelete(index: number){
+    if(!this.isAdmin){
+      return
+    }
     this.completedService.deleteTask(index);
   }
 }
